(1) New, Add Level Base Screen which is common and resuable for level 1 and 2
